Decouple HardwareWalletService into two classes: LedgerHardwareWalletService and HardwareWalletService
